Deprecate ContactDataService
The ContactDataService class and its methods have been marked as deprecated. This service was originally created to temporarily store contact data entered in the ContactCardComponent. While it is not currently in use, it has been kept in the codebase for potential future expansion or modification of how these data are handled.