Fix hydration issue from hero header component
Some common browser APIs and capabilities might not be available on the server. Applications cannot make use of browser-specific global objects like window, document, navigator, or location as well as certain properties of HTMLElement.

The particles-js should not be rendered in the server side. Use
hydration.

What is hydration? Hydration is the process that restores the server side rendered application on the client. This includes things like reusing the server rendered DOM structures, persisting the application state, transferring application data that was retrieved already by the server, and other processes.

Docs:
- https://angular.io/guide/ssr#hydration
- https://angular.io/guide/ssr#authoring-server-compatible-components