Add layout component in page not found routing
LayoutModule is imported in page-not-found.module.ts to ensure that PageNotFoundModule is self-contained and has direct access to the components and services it needs, such as LayoutComponent and its dependencies.

By making PageNotFoundModule self-sufficient, we adhere to the principles of modular design, facilitating better maintainability and reusability of the modules.